requirements.txt
# Only what's needed for Gemini integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=1.10.8,<2.0.0